use flask-login

+++ sgnotes_app/sgnotes_app.py
@@ -1,6 +1,7 @@
 from flask import Flask, render_template, session, redirect, url_for, request, flash
+from flask_login import login_user, logout_user, login_required
 from flask_sqlalchemy import SQLAlchemy
-from sgnotes_app import app, db
+from sgnotes_app import app, db, login_manager
 from .models import Note, User
 from .forms import NoteForm
 from sqlalchemy import desc
@@ -13,32 +14,30 @@ def main_page():
 
 
 @app.route('/', methods=['GET', 'POST'])
+@login_required
 def author_notes():
     """Главная страница пользователя. Содержит в себе поиск по методу POST.
     Сортирует по дате добавления.
     """
     user_id = session.get('user_id')
-    if user_id is None:
-        flash('Функция доступна только авторизованным пользователям')
-        return redirect(url_for('main_page'))
-    user = User.query.get(user_id)
-    notes = user.notes
+    user = load_user(user_id)
+    if user.notes is not None:
+        notes = user.notes
+    else:
+        flash('Заметок пока что - нет :(')
+        return render_template('author_notes.html')
     return render_template('author_notes.html', notes=notes)
-    # if 'username' in session:
-    #     username = session['username']
-    #     user_notes = Note.query.filter_by(user_nickname=username).all()
-    #     return render_template('index.html', username=username, notes=user_notes)
-    # else:
-    #     return redirect(url_for('register'))
 
 
 @app.route('/search', methods=['GET', 'POST'])
+@login_required
 def search():
     if request.method == 'POST':
         user_id = session.get('user_id')
+        user = load_user(user_id)
         if user_id:
             search_term = request.form['search_term']
-            notes = Note.query.filter(Note.title.ilike(f'%{search_term}%')).order_by(desc(Note.timestamp)).all()
+            notes = Note.query.filter(Note.title.ilike(f'%{search_term}%')).filter(Note.user_id==user).order_by(desc(Note.timestamp)).all()
             return render_template('author_notes.html', notes=notes)
         else:
             flash('Поиск по заметкам доступен только для авторизованных пользователей!')
@@ -46,18 +45,17 @@ def search():
 
 
 @app.route('/add-note', methods=['GET', 'POST'])
+@login_required
 def add_note():
     """Добавление заметки."""
     user_id = session.get('user_id')
-    if user_id is None:
-        flash('Функция доступна только авторизованным пользователям')
-        return redirect(url_for('main_page'))
+    user = load_user(user_id)
     form = NoteForm()
     if form.validate_on_submit():
         note = Note(
             title=form.title.data, 
             text=form.text.data,
-            user=user_id
+            user_id=user
         )
         if form.deadline.data:
             note.set_deadline(form.deadline.data)
@@ -68,48 +66,74 @@ def add_note():
 
 
 @app.route('/delete/<int:note_id>', methods=['GET', 'POST'])
+@login_required
 def delete_note(note_id):
     """Удаление заметки."""
+    user_id = session.get('user_id')
+    user = load_user(user_id)
     note = Note.query.get_or_404(note_id)
-    db.session.delete(note)
-    db.session.commit()
-    return redirect('/')
+    if note.user_id == user:
+        db.session.delete(note)
+        db.session.commit()
+        return redirect('/')
+    else:
+        return 'BANNED DEL!'
 
 
 @app.route('/edit/<int:note_id>', methods=['GET', 'POST'])
+@login_required
 def edit_note(note_id):
     """Редактирование заметки."""
+    user_id = session.get('user_id')
+    user = load_user(user_id)
     note = Note.query.get_or_404(note_id)
-    if request.method == 'POST':
-        note.title = request.form['title']
-        note.text = request.form['text']
-        if request.form['deadline']:
-            note.set_deadline(request.form['deadline'])
-        else:
-            note.deadline = None
-        db.session.commit()
-        return redirect(url_for('author_notes'))
-    return render_template('edit_notes.html', note=note)
+    if note.user_id == user:
+        if request.method == 'POST':
+            note.title = request.form['title']
+            note.text = request.form['text']
+            if request.form['deadline']:
+                note.set_deadline(request.form['deadline'])
+            else:
+                note.deadline = None
+            db.session.commit()
+            return redirect(url_for('author_notes'))
+        return render_template('edit_notes.html', note=note)
+    else:
+        return 'BANNED UPD!!!!!!'
 
 
 @app.route('/done/<int:note_id>', methods=['GET', 'POST'])
+@login_required
 def done_note(note_id):
     """Функция, которая отмечает задание выполненным."""
+    user_id = session.get('user_id')
+    user = load_user(user_id)
     note = Note.query.get_or_404(note_id)
-    note.is_done = not note.is_done
-    db.session.commit()
-    return redirect(request.referrer)
+    if note.user_id == user:
+        note.is_done = not note.is_done
+        db.session.commit()
+        return redirect(request.referrer)
+    else:
+        return 'BANNED DNNT!!!!!!!!!!'
 
 
 @app.route('/done/', methods=['GET'])
 @app.route('/undone/', methods=['GET'])
+@login_required
 def complete_task():
     """Страницы выполненных/невыполненных заданий."""
+    user_id = session.get('user_id')
+    user = load_user(user_id)
     if request.path == '/done/':
-        notes = Note.query.filter(Note.is_done==True).order_by(desc(Note.timestamp)).all()
+        notes = Note.query.filter(Note.is_done==True).filter(Note.user_id==user).order_by(desc(Note.timestamp)).all()
     elif request.path == '/undone/':
-        notes = Note.query.filter(Note.is_done==False).order_by(desc(Note.timestamp)).all()
+        notes = Note.query.filter(Note.is_done==False).filter(Note.user_id==user).order_by(desc(Note.timestamp)).all()
     return render_template('author_notes.html', notes=notes)
+
+
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -119,7 +143,7 @@ def login():
         password = request.form['password']
         user = User.query.filter_by(username=username).first()
         if user and user.password == password:
-            session['user_id'] = user.id
+            login_user(user)
             return redirect(url_for('author_notes'))
         return 'Неверные данные входа'
     return render_template('login.html')
@@ -136,6 +160,13 @@ def register():
         db.session.commit()
         return redirect(url_for('login'))
     return render_template('register.html')
+
+
+@app.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('login'))
 
 
 @app.route('/change_password', methods=['GET', 'POST'])
